Vulnerability fixes regarding manipulation of the API URL via the bot
Thanks to Lort533, this vulnerability was fixed. 
=====
SUMMARY OF THE VULNERABILITY REPORT SUBMITTED
--
Lort533 discovered a potential vulnerability in a bot's command system. This allows the bot to makes requests to an API endpoint to retrieve information about a specified username. The issue arises when a user sends a command with a username that includes special characters, such as "&". This causes the URL sent to the bot to be malformed, with the username parameter being escaped and allowing additional parameters to be appended.

He demonstrates that by sending a command like "k!name &another_parameter=value," the bot sends a request with a URL like "https://api.lixqa.de/v2/discord/pomelo-lookup/?username=&another_parameter=value." This is equivalent to "https://api.lixqa.de/v2/discord/pomelo-lookup/?another_parameter=value," effectively bypassing the intended username parameter. Although he failed to exploit the vulnerability by redirecting the bot to a malicious link, we were warned that a skilled hacker could potentially use this vulnerability to redirect the bot to a malicious website and potentially gain access to the hosting machine.

To mitigate this vulnerability, Lort533 suggests implementing code using regular expressions (regex) to validate and sanitize the username input. They provide an example of using the regex pattern "/^[_.A-Za-z0-9]+$/" (Line 28-31) to ensure that the username only contains alphanumeric characters, underscores, and dots. He also suggest checking the length of the user input to prevent excessively long usernames from causing issues.
==========
We have fixed this vulnerability, and we thank the community for their effort improving this open source bot project! 
==
CREDITS:
Lort533 (Vulnerability Reporter)
niksek11 (Updated the bot to Discord.js V14)